feat: Issue log with multiple columns changes.
