+++ src/main/java/org/spin/server/AllInOneServices.java
@@ -35,7 +35,6 @@ import org.spin.grpc.service.GeneralLedger;
 import org.spin.grpc.service.ImportFileLoader;
 import org.spin.grpc.service.InOutInfo;
 import org.spin.grpc.service.InvoiceInfo;
-import org.spin.grpc.service.IssueManagement;
 import org.spin.grpc.service.LogsInfo;
 import org.spin.grpc.service.MaterialManagement;
 import org.spin.grpc.service.NoticeManagement;
@@ -61,6 +60,7 @@ import org.spin.grpc.service.form.ExpressReceipt;
 import org.spin.grpc.service.form.ExpressShipment;
 import org.spin.grpc.service.form.PaymentAllocation;
 import org.spin.grpc.service.form.bank_statement_match.BankStatementMatch;
+import org.spin.grpc.service.form.issue_management.IssueManagement;
 import org.spin.grpc.service.form.match_po_receipt_invoice.MatchPOReceiptInvoice;
 import org.spin.grpc.service.form.payroll_action_notice.PayrollActionNotice;
 import org.spin.grpc.service.form.trial_balance_drillable.TrialBalanceDrillable;
