+++ src/main/proto/issue_management.proto
@@ -488,6 +488,13 @@ enum IssueCommentType {
 }
 
 // Issue Comment Definition
+message IssueCommentLog {
+	string column_name = 1;
+	string label = 2;
+	google.protobuf.Value new_value = 3;
+	string displayed_value = 4;
+}
+
 message IssueComment {
 	int32 id = 1;
 	string result = 2;
@@ -495,9 +502,7 @@ message IssueComment {
 	User user = 4;
 	IssueCommentType issue_comment_type = 5;
 	// log fields
-	string label = 6;
-	google.protobuf.Value new_value = 7;
-	string displayed_value = 8;
+	repeated IssueCommentLog change_logs = 6;
 }
 
 message ListIssueCommentsRequest {
